feat: 토론방 생성 api 연결

+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/network/request/CreateBookRequest.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CreateBookRequest(
-    val bookIsbn: Int,
+    val bookIsbn: Long,
     val bookTitle: String,
     val bookAuthor: String,
     val bookImage: String,
